Currently a little stuck...
I began implementing the plotting functions and to do so, I needed to make some variables global variables. That's fine, but for some reasons the masked arrays that I need to be global decide to change type to regular ndarrays. This is causing some issues with a ".filled" method I call at some point. I am pretty sure I can find an easy solution for this, but I just wanted to get this version up on the cloud before doing anything else.