Added back exkurt plotting and fixed command line usage... Maybe!
I added back the ability to plot excess kurtosis vs. frequency and I fixed the way command line usage works. Before, it wasn't truly taking all of the user-specifiable options, but it should be doing so now. I would test this, but COSMIC seems to be struggling a little bit at the moment. I will be back with updates at a later commit...