Back to a stable place...?
It seems that the tavg power plot is now working again... Not really sure what the issue was in the end? Something weird with log and having power = 0 messing things up it seems.

I'm going to continue making changes on this branch for a bit before merging back to oop_crickets and then to main.