Intermediate commit
So I fixed and then promptly broke the code again! I am making this commit because I now know what the issue is :) The problem has something to do with the "pows" and "freqs" variables and how they differ from the info_table columns that are supposed to be the same. I am not 100% sure what the issue is, but I do know that in places where I tried to eliminate redundancy, problems arose. Therefore, I will be reverting to the commit that I know is stable (5f1b758) and going from there. At this point, I will have to re-implement the plot_exkurt function, but that should not be too big of an issue.